⚰️ ♻️ 💥 Better integration/plotly code organize, + add required parameters, - removed tkinter useless interface
