--- integrations/plotly/README.md
@@ -1,34 +1,0 @@
-## QTSBE/integration/plotly
-
-The "display" module offers powerful features for visualizing and analyzing financial data using the Binance API and Plotly. With a user-friendly interface and interactive charts, users can explore market trends, evaluate trading strategy performance, and make informed investment decisions.
-
-### Usage
-
-To use the display module, follow these steps:
-
-1. **Select Data and Strategy**: Choose the data file to analyze and the trading strategy to evaluate using the dropdown menus available in the user interface.
-
-2. **Load Data**: Press the "Load" button to send a request to the Binance API and retrieve the necessary data based on the selected trading pair and strategy.
-
-3. **Analyzing Results**: Once the data is retrieved, a new window opens, displaying both the raw data in JSON format and an interactive graphical representation of the results using Plotly.
-
-### Features
-
-Here are the main features offered by the "display" module:
-
-- **Candlestick Charts**: Visualization of opening, closing, high, and low prices of an asset over a specified period.
-
-- **Technical Indicators**: Integration of indicators such as moving averages, relative strength index (RSI), moving average convergence/divergence (MACD), and Bollinger Bands.
-
-- **Trading Signals**: Graphical representation of buy and sell points generated by a specific trading strategy.
-
-- **Comparative Analysis**: Comparison of the performance of different assets or different strategies on the same chart.
-
-- **Transaction Analysis**: Visualization of transactions, such as buys and sells, to evaluate the performance of a trading strategy.
-
-- **Interactive Charts**: Interactive exploration of data with features such as zoom, scroll, and hover for detailed insights.
-
-- **Export and Share**: Exporting charts in various formats for integration into reports or presentations, as well as online sharing via URL links or integration into web applications.
-
-![image](https://github.com/simonpotel/QTSBE/blob/711bb2cecf12bdaef53d9d7a20f05e1971e4af59/files/display/python/full_view.png)
-![image](https://github.com/simonpotel/QTSBE/blob/06b8f029f0889afbb6ef6f95ca997f126c4c6f1c/files/display/python/price_extra_view.png)
